@@ -13,12 +13,15 @@
    "author": "",
    "license": "ISC",
    "devDependencies": {
+      "@types/app-root-path": "^1.2.4",
       "@types/enzyme": "^3.9.1",
       "@types/enzyme-adapter-react-16": "^1.0.5",
       "@types/jest": "^24.0.11",
-      "@types/react": "^16.8.8",
+      "@types/react": "^16.8.10",
       "@types/react-dom": "^16.8.3",
       "@types/selenium-webdriver": "^3.0.15",
+      "app-root-path": "^2.2.1",
+      "async-wait-until": "^1.2.4",
       "awesome-typescript-loader": "^5.2.1",
       "chromedriver": "^2.46.0",
       "clean-webpack-plugin": "^2.0.1",
@@ -28,6 +31,7 @@
       "geckodriver": "^1.16.0",
       "html-webpack-plugin": "^3.2.0",
       "identity-obj-proxy": "^3.0.0",
+      "import": "0.0.6",
       "jest": "^24.5.0",
       "less": "^3.9.0",
       "less-loader": "^4.1.0",
@@ -42,20 +46,22 @@
       "postcss-reporter": "^6.0.1",
       "postcss-url": "^8.0.0",
       "sass-loader": "^7.1.0",
+      "scub111-fileio": "^1.0.7",
       "selenium-webdriver": "^3.6.0",
       "style-loader": "^0.23.1",
-      "ts-jest": "^24.0.0",
+      "ts-jest": "^24.0.1",
       "ts-loader": "^5.3.3",
       "tslint": "^5.14.0",
       "tslint-config-airbnb": "^5.11.1",
-      "typescript": "^3.3.4000",
+      "typescript": "^3.4.1",
       "webpack": "^4.29.6",
       "webpack-cleanup-plugin": "^0.5.1",
       "webpack-cli": "^3.3.0",
       "webpack-dev-server": "^3.2.1"
    },
    "dependencies": {
-      "react": "^16.8.4",
-      "react-dom": "^16.8.4"
+      "axios": "^0.18.0",
+      "react": "^16.8.6",
+      "react-dom": "^16.8.6"
    }
 }
