@@ -1,14 +1,14 @@
 module.exports = {
    "roots": [
-      "<rootDir>"
+      "<rootDir>/src"
    ],
    "transform": {
       "^.+\\.tsx?$": "ts-jest"
    },
    //"testRegex": "(/__tests23__/.*|(\\.|/)(test|spec))\\.tsx?$",
    //"testRegex": "/__tests__/render/app/components/Posts/.*",
-   "testRegex": "/__tests__/auto/.*",
-   //"testRegex": "/__tests__/.*",
+   //"testRegex": "/__tests__/auto/.*",
+   "testRegex": "/__tests__/.*",
    "moduleFileExtensions": [
       "ts",
       "tsx",
@@ -18,6 +18,7 @@ module.exports = {
       "node"
    ],
    "moduleNameMapper": {
+      "src/(.*)": "<rootDir>/src/$1",
       "app/(.*)": "<rootDir>/src/app/$1",
       "\\.(css|less)$": "identity-obj-proxy"
    }
